requirements.txt
certifi==2024.8.30
charset-normalizer==3.3.2
feedparser==6.0.11
idna==3.10
newsapi-python==0.2.7
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
sgmllib3k==1.0.0
six==1.16.0
urllib3==2.2.3